Add relationships between businesses, users, employees, clients.
A user has many businesses (as owner).
A business belongs to a user (the owner).
A business has many employees.
A business has many clients.
A user has many employees (employee records).
A user has many clients (client records).
An employee belongs to a business.
An employee belongs to a user.
A client belongs to a business.
A client belongs to a user.